Add files via upload
The goal of this project the click conversation rate of the advertisements in a social media ap. App advertising plays a vital role in supporting smartphone apps. In order to maximize the expected revenue, the app must predict the probability of a click on an ad, more commonly known as click conversation rate (CVR) of an ad. The data set was a real world data sampled and organized from logs of clicks. The data includes detailed features related to ads, user, and context.